==> 03_encoding_categorical_numerical.R <==
library(tidyverse)
library(janitor)

# load household data created in 02_data_cleaning.R ----------------------------------------------------------------------

hhfile <- 
  read_csv("data/hh_merged_allyears.csv")

# Outcomes ---------------------------------------------------------------------------------------------------------------
hhfile <- 
  hhfile %>% 
  mutate(lnipcm = log(ipcm),
         totpov = factor(case_when(ipcm<linea_pobreza_total ~ "tot_pov",
                                  ipcm>=linea_pobreza_total ~ "nontot_pov"), 
         levels = c("tot_pov", "nontot_pov")),
         extpov = factor(case_when(ipcm<linea_pobreza_extrema ~ "ext_pov",
                                   ipcm>=linea_pobreza_extrema ~ "nonext_pov"), 
                         levels = c("ext_pov", "nonext_pov")))

# Predictors --------------------------------------------------------------------------------------------------------------
## factors ----------------------------------------------------------------------------------------------------------------
hhfile_clean <- 
  hhfile %>% 
  select(-c(vivi_vivienda, vivi_luz_electrica, vivi_linea_fija, vivi_banho, hh_personal_domestico, jefe_tiene_trabajo)) %>% # >98% in one category 
  mutate(area = as_factor(case_when(area==6 ~ "rural")),
         dptorep = as_factor(case_when(dptorep==2 ~ "San_Pedro",
                                       dptorep==5 ~ "Caaguazu",
                                       dptorep==6 ~ "Caazapa",
                                       dptorep==7 ~ "Itapua",
                                       dptorep==10 ~ "Alto_Parana",
                                       dptorep==11 ~ "Central",
                                       dptorep==20 ~ "Resto")),
         vivi_pared = as_factor(case_when(vivi_pared==1 ~ "Estaqueo",
                                          vivi_pared==2 ~ "Adobe",
                                          vivi_pared==3 ~ "Madera",
                                          vivi_pared==4 ~ "Ladrillo",
                                          vivi_pared==5 ~ "Cemento",
                                          vivi_pared==6 ~ "Palma",
                                          vivi_pared==7 ~ "Carton",
                                          vivi_pared==8 ~ "No_tiene",
                                          vivi_pared==9 ~ "Otro")),
         vivi_piso = as_factor(case_when(vivi_piso==1 ~ "Tierra",
                                         vivi_piso==2 ~ "Madera",
                                         vivi_piso==3 ~ "Ladrillo",
                                         vivi_piso==4 ~ "Lecherada",
                                         vivi_piso==5 ~ "Baldosa",
                                         vivi_piso==6 ~ "Porcelanato",
                                         vivi_piso==7 ~ "Parquet",
                                         vivi_piso==8 ~ "Alfombra",
                                         vivi_piso==9 ~ "Otro")),
         vivi_techo = as_factor(case_when(vivi_techo==1 ~ "Teja",
                                          vivi_techo==2 ~ "Paja",
                                          vivi_techo==3 ~ "Fibrocemento",
                                          vivi_techo==4 ~ "Zinc",
                                          vivi_techo==5 ~ "Madera",
                                          vivi_techo==6 ~ "Hormigon",
                                          vivi_techo==7 ~ "Palma",
                                          vivi_techo==8 ~ "Carton",
                                          vivi_techo==9 ~ "Otro")),
         vivi_agua_proveedor = as_factor(case_when(vivi_agua_proveedor==1 ~ "ESSAP",
                                                   vivi_agua_proveedor==2 ~ "SENASA",
                                                   vivi_agua_proveedor==3 ~ "Red_comunitaria",
                                                   vivi_agua_proveedor==4 ~ "Privada",
                                                   vivi_agua_proveedor==5 ~ "Artesiano",
                                                   vivi_agua_proveedor==6 ~ "Pozo_bomba",
                                                   vivi_agua_proveedor==7 ~ "Pozo_sin_bomba",
                                                   vivi_agua_proveedor==8 ~ "Manantial",
                                                   vivi_agua_proveedor==9 ~ "Rio",
                                                   vivi_agua_proveedor==10 ~ "Lluvia")),
         vivi_agua_fuente = as_factor(case_when(vivi_agua_fuente==1 ~ "Caneria_terreno",
                                                vivi_agua_fuente==2 ~ "Caneria_vivienda",
                                                vivi_agua_fuente==4 ~ "Pozo_terreno",
                                                vivi_agua_fuente==5 ~ "Vecino",
                                                vivi_agua_fuente==7 ~ "Otros")),
         vivi_agua_proveedor_beber = as_factor(case_when(vivi_agua_proveedor_beber==1 ~ "ESSAP",
                                                         vivi_agua_proveedor_beber==2 ~ "SENASA",
                                                         vivi_agua_proveedor_beber==3 ~ "Red_comunitaria",
                                                         vivi_agua_proveedor_beber==4 ~ "Privada",
                                                         vivi_agua_proveedor_beber==5 ~ "Artesiano",
                                                         vivi_agua_proveedor_beber==6 ~ "Pozo_protegido",
                                                         vivi_agua_proveedor_beber==7 ~ "Pozo_sin_proteccion",
                                                         vivi_agua_proveedor_beber==8 ~ "Manantial_protegido",
                                                         vivi_agua_proveedor_beber==9 ~ "Manantial_sin_proteccion",
                                                         vivi_agua_proveedor_beber==10 ~ "Lluvia",
                                                         vivi_agua_proveedor_beber==11 ~ "Embotellada",
                                                         vivi_agua_proveedor_beber==13 ~ "Rio")),
         vivi_agua_fuente_beber = as_factor(case_when(vivi_agua_fuente_beber==1 ~ "Caneria_terreno",
                                                      vivi_agua_fuente_beber==2 ~ "Canaeria_vivienda",
                                                      vivi_agua_fuente_beber==3 ~ "Canilla_publica",
                                                      vivi_agua_fuente_beber==4 ~ "Pozo_terreno",
                                                      vivi_agua_fuente_beber==5 ~ "Vecino",
                                                      vivi_agua_fuente_beber==7 ~ "Embotellada",
                                                      vivi_agua_fuente_beber==8 ~ "Otros",
                                                      vivi_agua_fuente_beber==9 ~ NA_character_)),
         vivi_banho_desague = as_factor(case_when(vivi_banho_desague==1 ~ "Red_sanitario",
                                                  vivi_banho_desague==2 ~ "Camara_septica",
                                                  vivi_banho_desague==3 ~ "Pozo_ciego",
                                                  vivi_banho_desague==4 ~ "Hoyo_abierto",
                                                  vivi_banho_desague==5 ~ "Letrina_ventilada",
                                                  vivi_banho_desague==6 ~ "Letrina_comun",
                                                  vivi_banho_desague==7 ~ "Letrina_comun_sin_techo",
                                                  vivi_banho_desague==8 ~ "Otro",
                                                  is.na(vivi_banho_desague) ~ "No_banho")),
         vivi_combustible = as_factor(case_when(vivi_combustible==1 ~ "Leña",
                                                vivi_combustible==2 ~ "Gas",
                                                vivi_combustible==3 ~ "Carbon",
                                                vivi_combustible==4 ~ "Electricidad",
                                                vivi_combustible==7 ~ "Ninguno")),
         vivi_basura = as_factor(case_when(vivi_basura==1 ~ "Quema",
                                           vivi_basura==2 ~ "Recoleccion_publica",
                                           vivi_basura==3 ~ "Recoleccion_privada",
                                           vivi_basura==4 ~ "Hoyo",
                                           vivi_basura==5 ~ "Patio",
                                           vivi_basura==6 ~ "Vertedero_municipal",
                                           vivi_basura==7 ~ "Chacra",
                                           vivi_basura==8 ~ "Arroyo",
                                           vivi_basura==9 ~ "Otro")),
         vivi_vivienda_propiedad = as_factor(case_when(vivi_vivienda_propiedad==1 ~ "Propia",
                                                       vivi_vivienda_propiedad==2 ~ "Cuotas",
                                                       vivi_vivienda_propiedad==3 ~ "Condo",
                                                       vivi_vivienda_propiedad==4 ~ "Alquilada",
                                                       vivi_vivienda_propiedad==5 ~ "Ocupada_Hecho",
                                                       vivi_vivienda_propiedad==6 ~ "Cedida")),
         vivi_lote_propiedad = as_factor(case_when(vivi_lote_propiedad==1 ~ "Propio",
                                                   vivi_lote_propiedad==2 ~ "Cuotas",
                                                   vivi_lote_propiedad==3 ~ "Condo",
                                                   vivi_lote_propiedad==4 ~ "Fiscal_municipal",
                                                   vivi_lote_propiedad==6 ~ "Ocupado_hecho",
                                                   vivi_lote_propiedad==7 ~ "Cedido",
                                                   is.na(vivi_lote_propiedad) ~ "No_propio")),
         hh_conyuge_present = as_factor(case_when(hh_conyuge_present==1 ~ "Yes",
                                                  hh_conyuge_present==0 ~ "No")),
         hh_tipo_hogar = as_factor(case_when(hh_tipo_hogar==1 ~ "Unipersonal",
                                             hh_tipo_hogar==2 ~ "Nuclear_Completo",
                                             hh_tipo_hogar==3 ~ "Nuclear_Incompleto",
                                             hh_tipo_hogar==4 ~ "Extendido",
                                             hh_tipo_hogar==5 ~ "Compuesto")),
         hh_tekopora = as_factor(case_when(hh_tekopora==1 ~ "Yes",
                                           hh_tekopora==0 ~ "No")),
         jefe_female = as_factor(case_when(jefe_female==1 ~ "Yes",
                                           jefe_female==0 ~ "No")),
         jefe_idioma = as_factor(case_when(jefe_idioma==1 ~ "Guarani",
                                           jefe_idioma==2 ~ "Bilingual",
                                           jefe_idioma==3 ~ "Castellano",
                                           jefe_idioma==4 ~ "Otro",
                                           jefe_idioma==5 ~ "No_habla")),
         jefe_estado_civil = as_factor(case_when(jefe_estado_civil==1 ~ "Casado",
                                                 jefe_estado_civil==2 ~ "Unido",
                                                 jefe_estado_civil==3 ~ "Separado",
                                                 jefe_estado_civil==4 ~ "Viudo",
                                                 jefe_estado_civil==5 ~ "Soltero",
                                                 jefe_estado_civil==6 ~ "Divorciado")),
         jefe_seguro_medico = as_factor(case_when(jefe_seguro_medico %in% c(1:6) ~ "Yes",
                                                  jefe_seguro_medico==7 ~ "No")),
         jefe_sabe_leer_escribir = as_factor(case_when(jefe_sabe_leer_escribir==1 ~ "Yes",
                                                       jefe_sabe_leer_escribir %in% c(2:6) ~ "No")),
         jefe_tipo_empleo = as_factor(case_when(jefe_tipo_empleo==1 ~ "Publico",
                                                jefe_tipo_empleo==2 ~ "Privado",
                                                jefe_tipo_empleo==3 ~ "Patron",
                                                jefe_tipo_empleo==4 ~ "Cuenta_propia",
                                                jefe_tipo_empleo==5 ~ "No_remunerado",
                                                jefe_tipo_empleo==6 ~ "Domestico",
                                                jefe_tipo_empleo==7 ~ "Empleado_extranjero",
                                                jefe_tipo_empleo==8 ~ "Empleador_extranjero",
                                                jefe_tipo_empleo==99 ~ "No_trabajo")),
         jefe_quiere_trabajo = as_factor(case_when(jefe_quiere_trabajo==1 ~ "Yes",
                                                   jefe_quiere_trabajo==0 ~ "No")),
         jefe_caja_jubilacion1 = as_factor(case_when(jefe_caja_jubilacion1==1 ~ "Yes",
                                                     jefe_caja_jubilacion1 %in% c(0, 6) ~ "No"))) %>% 
  mutate_at(vars(vivi_celular, vivi_pieza_cocinar, vivi_computadora, vivi_tableta, vivi_internet, vivi_radio, vivi_televisor, vivi_heladera,
                 vivi_cocina_gas, vivi_cocina_elec, vivi_lavarropas, vivi_videoDVD, vivi_termocalefon, vivi_acondicionador_aire,
                 vivi_antena_parabolica, vivi_tv_cable, vivi_horno_microondas, vivi_horno_electrico, vivi_auto_camion, vivi_motocicleta),
            ~as_factor(case_when(.==1 ~ "Yes",
                                 .==6 ~ "No",
                                 .==9 ~ NA_character_)))

## numeric -----------------------------------------------------------------------------------------------
hhfile_clean <- 
  hhfile_clean %>% 
  mutate(vivi_piezas = na_if(vivi_piezas, 99),
         vivi_dormitorios = na_if(vivi_dormitorios, 99),
         jefe_aniosestudio = na_if(jefe_aniosestudio, 99))


#save data as rds to keep structure ------------------------------------------------------------------------
write_rds(hhfile_clean, "data/hh_merged_allyears_cleaned.rds")

# renv ---------------------------------------------------------------------------------------
renv::snapshot()
